Put createImage mutation in common declaration file
